More code cleanup
Address some compiler and code analyzer warnings.

+++ cx/platform/win/win_debug_crash.c
@@ -417,7 +417,7 @@ bool dbgCrashSetPath(strref path)
     crashdir = cstrDupw(tmp);
 
     // precompute filename we're going to use
-    processname = xaAlloc(strLen(exename) + 1);
+    processname = xaAlloc((size_t)strLen(exename) + 1);
     strCopyOut(exename, 0, processname, strLen(exename) + 1);
     processnamew = strToUTF16A(exename);
